test(tn-reth): move lib.rs unit tests into their implementation modules
